fix(backend): 修改monitor流程
