+++ src/Test/dummyRM.py
@@ -1,0 +1,6 @@
+# coding=utf-8
+""" docstring: 简易RM模拟器 """
+
+
+if __name__ == "__main__":
+    pass
